+++ index.php
@@ -43,10 +43,6 @@ require_once(dirname(__FILE__) . '/Settings.php');
 if ((empty($cachedir) || !file_exists($cachedir)) && file_exists($boarddir . '/cache'))
 	$cachedir = $boarddir . '/cache';
 
-// For php below 7
-if (!function_exists('random_int'))
-	require_once ($sourcedir . '/random_compat/random_int.php');
-
 // Without those we can't go anywhere
 require_once($sourcedir . '/QueryString.php');
 require_once($sourcedir . '/Subs.php');
